+++ OSDiagTool/OSDiagTool.csproj
@@ -124,8 +124,9 @@
     <Compile Include="Platform\ConfigFiles\ConfigFileDBInfo.cs" />
     <Compile Include="Platform\ConfigFiles\ConfigFileProperty.cs" />
     <Compile Include="Platform\ConfigFiles\ConfigFileReader.cs" />
-    <Compile Include="OSServiceConfigFileParser.cs" />
+    <Compile Include="Platform\OSServiceConfigFileParser.cs" />
     <Compile Include="Platform\LogExporter.cs" />
+    <Compile Include="Platform\PlatformFilesHelper.cs" />
     <Compile Include="Platform\PlatformVersion.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
