+++ OSDiagTool/OSDiagToolForm/PopUpForm.cs
@@ -20,25 +20,36 @@ namespace OSDiagTool.OSDiagToolForm {
             InitializeComponent();
 
             if (feedbackType.Equals(_feedbackWaitType)) {
+
+                lb_ProgressFeedback.Visible = true;
+                lb_ProgressFeedback.Width = lb_ProgressFeedback.Width * 5 / 4;
+                lb_ProgressFeedback.Height = lb_ProgressFeedback.Height * 4;
+                this.Height = this.Height * 3 / 2;
+                this.Width = this.Width * 5 / 4;
+
                 bt_CloseFormPopUp.Visible = false;
-                this.Width = this.Width * 6/5;
                 pb_progressBar.Style = ProgressBarStyle.Continuous;
-                pb_progressBar.Width = pb_progressBar.Width * 6 / 5;
+                pb_progressBar.Width = pb_progressBar.Width * 5 / 4;
                 pb_progressBar.Maximum = 100;
                 pb_progressBar.Value = 0;
                 pb_progressBar.Step = pb_progressBar.Maximum/totalSteps;
                 pb_progressBar.PerformStep();
                 pb_progressBar.Visible = true;
+                pb_progressBar.Location = new Point(pb_progressBar.Location.X, pb_progressBar.Location.Y * 3);
 
             } else if (feedbackType.Equals(_feedbackDoneType)) {
                 this.Width = this.Width * 3;
                 bt_CloseFormPopUp.Location = new Point(bt_CloseFormPopUp.Location.X * 4, bt_CloseFormPopUp.Location.Y);
+                this.lbl_message.Text = message;
+
             } else if (feedbackType.Equals(_feedbackErrorType)) {
                 bt_CloseFormPopUp.Visible = true;
                 bt_CloseFormPopUp.Text = "Exit";
-            }
+                this.lbl_message.Text = message;
 
-            this.lbl_message.Text = message;
+            }else if (feedbackType.Equals(_feedbackTestConnectionType)) {
+                this.lbl_message.Text = message;
+            }
 
         }
 
@@ -49,8 +60,11 @@ namespace OSDiagTool.OSDiagToolForm {
 
         public static void ChangeFeedbackLabelAndProgressBar (puf_popUpForm popUpForm, string labelText) {
 
-            popUpForm.lbl_message.Text = labelText;
-            if(!(popUpForm.pb_progressBar.Value >= popUpForm.pb_progressBar.Maximum)) {
+            popUpForm.lb_ProgressFeedback.Items.Add(labelText);
+            popUpForm.lb_ProgressFeedback.SelectedIndex = popUpForm.lb_ProgressFeedback.Items.Count - 1;
+            popUpForm.lb_ProgressFeedback.SelectedIndex = -1;
+
+            if (!(popUpForm.pb_progressBar.Value >= popUpForm.pb_progressBar.Maximum)) {
                 popUpForm.pb_progressBar.PerformStep();
             }
